refactor: move listModels into AgentPlugin; remove if/else ladder from modes.ts
AgentPlugin interface gains listModels() — each plugin owns its own
model discovery strategy:
  claude.ts  → returns hardcoded curated list (no subprocess)
  opencode.ts → shells out to `opencode models`
  cursor.ts  → shells out to `cursor-agent --list-models`

modes.ts fetchCliModelsUncached now delegates unconditionally:
  resolvePlugin(cli).listModels()

Removed: CLAUDE_MODELS_HARDCODED constant, execFile/promisify imports,
and the claude fast-path in resolveCliModels (unified cache path handles
all CLIs identically now).

AGENTS.md: document the plugin invariant — all CLI-specific logic belongs
in the plugin, never as if/else on CliId in calling code.
